back-end/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-mapbox-baselayer==1.0.0
djangorestframework==3.15.1
idna==3.6
oauthlib==3.2.2
psycopg==3.1.18
psycopg-binary==3.1.18
pygame==2.5.2
python-decouple==3.8
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
